requirements.txt
allure_python_commons==2.13.5
allure_pytest==2.13.5
Jinja2==3.1.5
jsonpath==0.82.2
openpyxl==3.1.5
PyMySQL==1.1.1
pytest==8.3.4
Requests==2.31.0